src/trainer/requirements.txt
coremltools  # ~=5.2
# this is only temporary solution (version 3.20+ breaks protobuf):
protobuf~=3.19.0
fastparquet==0.7.2
# consider pinning xgb version if improveai's xgboost version will be bumped to most recent
# xgboost==1.4.2
dask
dask_ml
pandas~=1.3.4
s3fs==0.4.2
# s3fs~=2022.3.0
Flask==2.0.2
gevent==21.8.0
gunicorn==20.1.0
# explicit orjson and ksuid specification although improveai also depends on them
orjson
svix-ksuid
# coremltools limits max usable version to 1.4.2 -> check once in  a while
xgboost==1.4.2
xxhash
